scripts/bonito/requirements.txt
# general requirements
mappy==2.24
toml==0.10.2
tqdm==4.31.1
scipy==1.5.4
numpy~=1.19.5
pysam==0.18.0
parasail==1.2
pandas==1.1.5
requests==2.25.1
ont-koi==0.0.7
ont-remora==0.1.2
ont-fast5-api==3.3.0
fast-ctc-decode==0.3.0
python-dateutil==2.8.2
# cuda requirements
#torch==1.10.0+cu113
